speech_translations/requirements.txt
azure-cognitiveservices-speech==1.30.0
azure-ai-textanalytics==5.3.0
setuptools
dotenv